// src/API.ts
export type Admin = {
  __typename: "Admin",
  address?: string | null,
  assignedParkingIds?: Array< string | null > | null,
  buildingCode: string,
  buildingName?: string | null,
  buildingNo?: string | null,
  cognitoUsername?: string | null,
  createdAt: string,
  email: string,
  id: string,
  managerCode: string,
  managerName: string,
  phoneNo?: string | null,
  updatedAt: string,
};

export type Building = {
  __typename: "Building",
  address: string,
  buildingCode: string,
  buildingName: string,
  buildingNo: string,
  createdAt: string,
  id: string,
  location?: string | null,
  updatedAt: string,
};

export type Parking = {
  __typename: "Parking",
  buildingCode: string,
  buildingName?: string | null,
  buildingNo?: string | null,
  createdAt: string,
  description?: string | null,
  id: string,
  parkingLots?: number | null,
  parkingName?: string | null,
  parkingNo: string,
  updatedAt: string,
};

export type ParkingSpot = {
  __typename: "ParkingSpot",
  buildingCode: string,
  createdAt: string,
  description?: string | null,
  floor?: string | null,
  id: string,
  isAvailable?: boolean | null,
  parkingId: string,
  section?: string | null,
  spotNumber: string,
  spotType?: ParkingSpotSpotType | null,
  updatedAt: string,
};

export enum ParkingSpotSpotType {
  ELECTRIC = "ELECTRIC",
  GUEST = "GUEST",
  REGULAR = "REGULAR",
}


export type Reserving = {
  __typename: "Reserving",
  accessNo: string,
  buildingCode: string,
  createdAt: string,
  email?: string | null,
  endTime: string,
  id: string,
  phone?: string | null,
  reservedBy?: string | null,
  spotId: string,
  startTime: string,
  status?: ReservingStatus | null,
  updatedAt: string,
  vehicleCode?: string | null,
};

export enum ReservingStatus {
  ACTIVE = "ACTIVE",
  CANCELLED = "CANCELLED",
  COMPLETED = "COMPLETED",
}


export type UnitInfo = {
  __typename: "UnitInfo",
  accessNo: string,
  buildingCode: string,
  cognitoUsername?: string | null,
  createdAt: string,
  email?: string | null,
  firstName: string,
  id: string,
  lastName: string,
  phone?: string | null,
  sakenLastName?: string | null,
  sakenName?: string | null,
  unitNo: string,
  updatedAt: string,
  username: string,
};

export type ModelAdminFilterInput = {
  address?: ModelStringInput | null,
  and?: Array< ModelAdminFilterInput | null > | null,
  assignedParkingIds?: ModelStringInput | null,
  buildingCode?: ModelStringInput | null,
  buildingName?: ModelStringInput | null,
  buildingNo?: ModelStringInput | null,
  cognitoUsername?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  email?: ModelStringInput | null,
  id?: ModelIDInput | null,
  managerCode?: ModelStringInput | null,
  managerName?: ModelStringInput | null,
  not?: ModelAdminFilterInput | null,
  or?: Array< ModelAdminFilterInput | null > | null,
  phoneNo?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};
// ...
export type ModelAdminConnection = {
  __typename: "ModelAdminConnection",
  items:  Array<Admin | null >,
  nextToken?: string | null,
};

export type ModelBuildingFilterInput = {
  address?: ModelStringInput | null,
  and?: Array< ModelBuildingFilterInput | null > | null,
  buildingCode?: ModelStringInput | null,
  buildingName?: ModelStringInput | null,
  buildingNo?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  id?: ModelIDInput | null,
  location?: ModelStringInput | null,
  not?: ModelBuildingFilterInput | null,
  or?: Array< ModelBuildingFilterInput | null > | null,
  updatedAt?: ModelStringInput | null,
};

export type ModelBuildingConnection = {
  __typename: "ModelBuildingConnection",
  items:  Array<Building | null >,
  nextToken?: string | null,
};

export type ListCognitoUsersReturnType = {
  __typename: "ListCognitoUsersReturnType",
  items:  Array<CognitoUser | null >,
  nextToken?: string | null,
};

export type CognitoUser = {
  __typename: "CognitoUser",
  accessNo?: string | null,
  buildingCode?: string | null,
  createdAt?: string | null,
  email: string,
  groups?: Array< string | null > | null,
  id: string,
  isActive?: boolean | null,
  managerCode?: string | null,
  name?: string | null,
  phone?: string | null,
  unitNo?: string | null,
  updatedAt?: string | null,
  userRole?: string | null,
  userStatus?: string | null,
};
// ...
export type GetAdminQueryVariables = {
  id: string,
};

export type GetAdminQuery = {
  getAdmin?:  {
    __typename: "Admin",
    address?: string | null,
    assignedParkingIds?: Array< string | null > | null,
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    cognitoUsername?: string | null,
    createdAt: string,
    email: string,
    id: string,
    managerCode: string,
    managerName: string,
    phoneNo?: string | null,
    updatedAt: string,
  } | null,
};

export type GetBuildingQueryVariables = {
  id: string,
};

export type GetBuildingQuery = {
  getBuilding?:  {
    __typename: "Building",
    address: string,
    buildingCode: string,
    buildingName: string,
    buildingNo: string,
    createdAt: string,
    id: string,
    location?: string | null,
    updatedAt: string,
  } | null,
};

export type GetParkingQueryVariables = {
  id: string,
};

export type GetParkingQuery = {
  getParking?:  {
    __typename: "Parking",
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    createdAt: string,
    description?: string | null,
    id: string,
    parkingLots?: number | null,
    parkingName?: string | null,
    parkingNo: string,
    updatedAt: string,
  } | null,
};

export type GetParkingSpotQueryVariables = {
  id: string,
};

export type GetParkingSpotQuery = {
  getParkingSpot?:  {
    __typename: "ParkingSpot",
    buildingCode: string,
    createdAt: string,
    description?: string | null,
    floor?: string | null,
    id: string,
    isAvailable?: boolean | null,
    parkingId: string,
    section?: string | null,
    spotNumber: string,
    spotType?: ParkingSpotSpotType | null,
    updatedAt: string,
  } | null,
};

export type GetReservingQueryVariables = {
  id: string,
};

export type GetReservingQuery = {
  getReserving?:  {
    __typename: "Reserving",
    accessNo: string,
    buildingCode: string,
    createdAt: string,
    email?: string | null,
    endTime: string,
    id: string,
    phone?: string | null,
    reservedBy?: string | null,
    spotId: string,
    startTime: string,
    status?: ReservingStatus | null,
    updatedAt: string,
    vehicleCode?: string | null,
  } | null,
};

export type GetUnitInfoQueryVariables = {
  id: string,
};

export type GetUnitInfoQuery = {
  getUnitInfo?:  {
    __typename: "UnitInfo",
    accessNo: string,
    buildingCode: string,
    cognitoUsername?: string | null,
    createdAt: string,
    email?: string | null,
    firstName: string,
    id: string,
    lastName: string,
    phone?: string | null,
    sakenLastName?: string | null,
    sakenName?: string | null,
    unitNo: string,
    updatedAt: string,
    username: string,
  } | null,
};

export type ListAdminsQueryVariables = {
  filter?: ModelAdminFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListAdminsQuery = {
  listAdmins?:  {
    __typename: "ModelAdminConnection",
    items:  Array< {
      __typename: "Admin",
      address?: string | null,
      assignedParkingIds?: Array< string | null > | null,
      buildingCode: string,
      buildingName?: string | null,
      buildingNo?: string | null,
      cognitoUsername?: string | null,
      createdAt: string,
      email: string,
      id: string,
      managerCode: string,
      managerName: string,
      phoneNo?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListBuildingsQueryVariables = {
  filter?: ModelBuildingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListBuildingsQuery = {
  listBuildings?:  {
    __typename: "ModelBuildingConnection",
    items:  Array< {
      __typename: "Building",
      address: string,
      buildingCode: string,
      buildingName: string,
      buildingNo: string,
      createdAt: string,
      id: string,
      location?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListCognitoUsersQueryVariables = {
  filter?: string | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListCognitoUsersQuery = {
  listCognitoUsers?:  {
    __typename: "ListCognitoUsersReturnType",
    items:  Array< {
      __typename: "CognitoUser",
      accessNo?: string | null,
      buildingCode?: string | null,
      createdAt?: string | null,
      email: string,
      groups?: Array< string | null > | null,
      id: string,
      isActive?: boolean | null,
      managerCode?: string | null,
      name?: string | null,
      phone?: string | null,
      unitNo?: string | null,
      updatedAt?: string | null,
      userRole?: string | null,
      userStatus?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListParkingSpotsQueryVariables = {
  filter?: ModelParkingSpotFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListParkingSpotsQuery = {
  listParkingSpots?:  {
    __typename: "ModelParkingSpotConnection",
    items:  Array< {
      __typename: "ParkingSpot",
      buildingCode: string,
      createdAt: string,
      description?: string | null,
      floor?: string | null,
      id: string,
      isAvailable?: boolean | null,
      parkingId: string,
      section?: string | null,
      spotNumber: string,
      spotType?: ParkingSpotSpotType | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListParkingsQueryVariables = {
  filter?: ModelParkingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListParkingsQuery = {
  listParkings?:  {
    __typename: "ModelParkingConnection",
    items:  Array< {
      __typename: "Parking",
      buildingCode: string,
      buildingName?: string | null,
      buildingNo?: string | null,
      createdAt: string,
      description?: string | null,
      id: string,
      parkingLots?: number | null,
      parkingName?: string | null,
      parkingNo: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListReservingsQueryVariables = {
  filter?: ModelReservingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListReservingsQuery = {
  listReservings?:  {
    __typename: "ModelReservingConnection",
    items:  Array< {
      __typename: "Reserving",
      accessNo: string,
      buildingCode: string,
      createdAt: string,
      email?: string | null,
      endTime: string,
      id: string,
      phone?: string | null,
      reservedBy?: string | null,
      spotId: string,
      startTime: string,
      status?: ReservingStatus | null,
      updatedAt: string,
      vehicleCode?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListUnitInfosQueryVariables = {
  filter?: ModelUnitInfoFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListUnitInfosQuery = {
  listUnitInfos?:  {
    __typename: "ModelUnitInfoConnection",
    items:  Array< {
      __typename: "UnitInfo",
      accessNo: string,
      buildingCode: string,
      cognitoUsername?: string | null,
      createdAt: string,
      email?: string | null,
      firstName: string,
      id: string,
      lastName: string,
      phone?: string | null,
      sakenLastName?: string | null,
      sakenName?: string | null,
      unitNo: string,
      updatedAt: string,
      username: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};